Add some tests with fixtures
