+++ package.json
@@ -4,7 +4,7 @@
   "description": "Convert Rework style CSS to SCSS",
   "main": "index.js",
   "scripts": {
-    "test": "node test"
+    "test": "mocha test"
   },
   "author": "Brent Jackson",
   "license": "MIT",
@@ -13,6 +13,7 @@
     "postcss": "^3.0.7"
   },
   "devDependencies": {
+    "mocha": "^2.1.0",
     "node-sass": "^1.2.3",
     "rework": "^1.0.1",
     "rework-npm": "^1.0.0"
