+++ test/fixtures/basscss-base.css
@@ -1,0 +1,248 @@
+/*
+    Basscss Base
+*/
+
+:root {
+  --font-family: 'Helvetica Neue', Helvetica, sans-serif;
+  --line-height: 1.5;
+  --heading-font-family: var(--font-family);
+  --heading-font-weight: bold;
+  --heading-line-height: 1.25;
+  --monospace-font-family: 'Source Code Pro', Consolas, monospace;
+  --h1: 2rem;
+  --h2: 1.5rem;
+  --h3: 1.25rem;
+  --h4: 1rem;
+  --h5: .875rem;
+  --h6: .75rem;
+  --bold-font-weight: bold;
+  --space-1: .5rem;
+  --space-2: 1rem;
+  --space-3: 2rem;
+  --space-4: 4rem;
+}
+
+body,
+button {
+  margin: 0;
+  padding: calc(2 * var(--space-2));
+}
+
+button,
+input,
+select,
+textarea {
+  font-family: inherit;
+  font-size: 100%;
+}
+
+img {
+  max-width: 100%;
+}
+
+svg {
+  max-height: 100%;
+}
+
+body {
+  font-family: var(--font-family);
+  line-height: var(--line-height);
+  font-size: 100%;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-family: var(--heading-font-family);
+  font-weight: var(--heading-font-weight);
+  line-height: var(--heading-line-height);
+  margin-top: 1em;
+  margin-bottom: .5em;
+}
+
+p,
+dl,
+ol,
+ul {
+  font-size: var(--h4);
+  margin-top: 0;
+  margin-bottom: var(--space-2);
+}
+
+ol,
+ul {
+  padding-left: var(--space-3);
+}
+
+pre,
+code,
+samp {
+  font-family: var(--monospace-font-family);
+  font-size: inherit;
+}
+
+pre {
+  margin-top: 0;
+  margin-bottom: var(--space-2);
+  overflow-x: scroll;
+}
+
+hr {
+  margin-top: var(--space-3);
+  margin-bottom: var(--space-3);
+}
+
+blockquote {
+  margin-top: var(--space-3);
+  margin-bottom: var(--space-3);
+  margin-left: 0;
+  padding-left: var(--space-2);
+  padding-right: var(--space-2);
+}
+
+blockquote,
+blockquote p {
+  font-size: var(--h3);
+  font-style: italic;
+}
+
+h1,
+.h1 {
+  font-size: var(--h1);
+}
+
+h2,
+.h2 {
+  font-size: var(--h2);
+}
+
+h3,
+.h3 {
+  font-size: var(--h3);
+}
+
+h4,
+.h4 {
+  font-size: var(--h4);
+}
+
+h5,
+.h5 {
+  font-size: var(--h5);
+}
+
+h6,
+.h6 {
+  font-size: var(--h6);
+}
+
+.list-reset {
+  list-style: none;
+  padding-left: 0;
+}
+
+input,
+select,
+textarea,
+fieldset {
+  margin-top: 0;
+  margin-bottom: var(--space-1);
+}
+
+input[type=text],
+input[type=datetime],
+input[type=datetime-local],
+input[type=email],
+input[type=month],
+input[type=number],
+input[type=password],
+input[type=search],
+input[type=tel],
+input[type=time],
+input[type=url],
+input[type=week] {
+  box-sizing: border-box;
+  height: 2.25em;
+  padding: .25em var(--space-2);
+  -webkit-appearance: none;
+}
+
+select {
+  box-sizing: border-box;
+  line-height: 1.75;
+  padding: .25em var(--space-2);
+}
+
+select:not([multiple]) {
+  height: 2.25em;
+}
+
+textarea {
+  box-sizing: border-box;
+  line-height: 1.75;
+  padding: .25em var(--space-2);
+}
+
+.fieldset-reset {
+  padding: 0;
+  margin-left: 0;
+  margin-right: 0;
+  border: 0;
+}
+
+.fieldset-reset legend {
+  padding: 0;
+}
+
+button,
+.button {
+  font-weight: var(--bold-font-weight);
+  text-decoration: none;
+  cursor: pointer;
+  display: inline-block;
+  box-sizing: border-box;
+  line-height: 1.125;
+  padding: .5em var(--space-2);
+  margin: 0;
+  height: auto;
+  border: 1px solid transparent;
+  -webkit-appearance: none;
+}
+
+::-moz-focus-inner {
+  border: 0;
+  padding: 0;
+}
+
+.button:hover {
+  text-decoration: none;
+}
+
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
+  max-width: 100%;
+  width: 100%;
+}
+
+th {
+  text-align: left;
+  font-weight: var(--bold-font-weight);
+}
+
+th,
+td {
+  padding: .25em var(--space-2);
+  line-height: inherit;
+}
+
+th {
+  vertical-align: bottom;
+}
+
+td {
+  vertical-align: top;
+}
