+++ test/fixtures/basscss-base.expected.scss
@@ -1,32 +1,35 @@
 
+// Converted Variables
+
+$font-family: 'Helvetica Neue', Helvetica, sans-serif !default;
+$line-height: 1.5 !default;
+$heading-font-family: $font-family !default;
+$heading-font-weight: bold !default;
+$heading-line-height: 1.25 !default;
+$monospace-font-family: 'Source Code Pro', Consolas, monospace !default;
+$h1: 2rem !default;
+$h2: 1.5rem !default;
+$h3: 1.25rem !default;
+$h4: 1rem !default;
+$h5: .875rem !default;
+$h6: .75rem !default;
+$bold-font-weight: bold !default;
+$space-1: .5rem !default;
+$space-2: 1rem !default;
+$space-3: 2rem !default;
+$space-4: 4rem !default;
+
+// Custom Media Query Variables
+
+
 /*
     Basscss Base
 */
 
-:root {
-  --font-family: 'Helvetica Neue', Helvetica, sans-serif;
-  --line-height: 1.5;
-  --heading-font-family: var(--font-family);
-  --heading-font-weight: bold;
-  --heading-line-height: 1.25;
-  --monospace-font-family: 'Source Code Pro', Consolas, monospace;
-  --h1: 2rem;
-  --h2: 1.5rem;
-  --h3: 1.25rem;
-  --h4: 1rem;
-  --h5: .875rem;
-  --h6: .75rem;
-  --bold-font-weight: bold;
-  --space-1: .5rem;
-  --space-2: 1rem;
-  --space-3: 2rem;
-  --space-4: 4rem;
-}
-
 body,
 button {
   margin: 0;
-  padding: calc(2 * var(--space-2));
+  padding: (2 * $space-2);
 }
 
 button,
@@ -46,8 +49,8 @@ svg {
 }
 
 body {
-  font-family: var(--font-family);
-  line-height: var(--line-height);
+  font-family: $font-family;
+  line-height: $line-height;
   font-size: 100%;
 }
 
@@ -57,9 +60,9 @@ h3,
 h4,
 h5,
 h6 {
-  font-family: var(--heading-font-family);
-  font-weight: var(--heading-font-weight);
-  line-height: var(--heading-line-height);
+  font-family: $heading-font-family;
+  font-weight: $heading-font-weight;
+  line-height: $heading-line-height;
   margin-top: 1em;
   margin-bottom: .5em;
 }
@@ -68,76 +71,76 @@ p,
 dl,
 ol,
 ul {
-  font-size: var(--h4);
+  font-size: $h4;
   margin-top: 0;
-  margin-bottom: var(--space-2);
+  margin-bottom: $space-2;
 }
 
 ol,
 ul {
-  padding-left: var(--space-3);
+  padding-left: $space-3;
 }
 
 pre,
 code,
 samp {
-  font-family: var(--monospace-font-family);
+  font-family: $monospace-font-family;
   font-size: inherit;
 }
 
 pre {
   margin-top: 0;
-  margin-bottom: var(--space-2);
+  margin-bottom: $space-2;
   overflow-x: scroll;
 }
 
 hr {
-  margin-top: var(--space-3);
-  margin-bottom: var(--space-3);
+  margin-top: $space-3;
+  margin-bottom: $space-3;
 }
 
 blockquote {
-  margin-top: var(--space-3);
-  margin-bottom: var(--space-3);
+  margin-top: $space-3;
+  margin-bottom: $space-3;
   margin-left: 0;
-  padding-left: var(--space-2);
-  padding-right: var(--space-2);
+  padding-left: $space-2;
+  padding-right: $space-2;
 }
 
 blockquote,
 blockquote p {
-  font-size: var(--h3);
+  font-size: $h3;
   font-style: italic;
 }
 
 h1,
 .h1 {
-  font-size: var(--h1);
+  font-size: $h1;
 }
 
 h2,
 .h2 {
-  font-size: var(--h2);
+  font-size: $h2;
 }
 
 h3,
 .h3 {
-  font-size: var(--h3);
+  font-size: $h3;
 }
 
 h4,
 .h4 {
-  font-size: var(--h4);
+  font-size: $h4;
 }
 
 h5,
 .h5 {
-  font-size: var(--h5);
+  font-size: $h5;
 }
 
 h6,
 .h6 {
-  font-size: var(--h6);
+  font-size: $h6;
 }
 
 .list-reset {
@@ -150,7 +153,7 @@ select,
 textarea,
 fieldset {
   margin-top: 0;
-  margin-bottom: var(--space-1);
+  margin-bottom: $space-1;
 }
 
 input[type=text],
@@ -167,14 +170,14 @@ input[type=url],
 input[type=week] {
   box-sizing: border-box;
   height: 2.25em;
-  padding: .25em var(--space-2);
+  padding: .25em $space-2;
   -webkit-appearance: none;
 }
 
 select {
   box-sizing: border-box;
   line-height: 1.75;
-  padding: .25em var(--space-2);
+  padding: .25em $space-2;
 }
 
 select:not([multiple]) {
@@ -184,7 +187,7 @@ select:not([multiple]) {
 textarea {
   box-sizing: border-box;
   line-height: 1.75;
-  padding: .25em var(--space-2);
+  padding: .25em $space-2;
 }
 
 .fieldset-reset {
@@ -200,13 +203,13 @@ textarea {
 
 button,
 .button {
-  font-weight: var(--bold-font-weight);
+  font-weight: $bold-font-weight;
   text-decoration: none;
   cursor: pointer;
   display: inline-block;
   box-sizing: border-box;
   line-height: 1.125;
-  padding: .5em var(--space-2);
+  padding: .5em $space-2;
   margin: 0;
   height: auto;
   border: 1px solid transparent;
@@ -231,12 +234,12 @@ table {
 
 th {
   text-align: left;
-  font-weight: var(--bold-font-weight);
+  font-weight: $bold-font-weight;
 }
 
 th,
 td {
-  padding: .25em var(--space-2);
+  padding: .25em $space-2;
   line-height: inherit;
 }
