+++ test/fixtures/basscss-base.output.scss
@@ -1,4 +1,3 @@
-
 // Converted Variables
 
 $font-family: 'Helvetica Neue', Helvetica, sans-serif !default;
@@ -20,7 +19,6 @@ $space-3: 2rem !default;
 $space-4: 4rem !default;
 
 // Custom Media Query Variables
-
 
 
 /*
